LIU-402: Complete separation of logical graph tests.
