+++ daliuge-translator/test/dropmake/test_lg.py
@@ -24,48 +24,47 @@ import json
 import pickle
 import unittest
 
-try:
-    from importlib.resources import files, as_file
-except (ImportError, ModuleNotFoundError):
-    from importlib_resources import files
-
 from dlg.common import CategoryType
 from dlg.dropmake import path_utils
 from dlg.dropmake.lg import LG
-
 
 NODES = 'nodeDataArray'
 LINKS = 'linkDataArray'
 TEST_SSID = 'test_pg_gen'  # Needed to match output files generated in test_pg_gen.py
 
 
-class TestLGNodeInit(unittest.TestCase):
+class TestLGInit(unittest.TestCase):
+    lg_names = {
+        "HelloWorld_simple.graph": 2,
+        "eagle_gather_empty_update.graph": 11,
+        "eagle_gather_simple_update.graph": 18,
+        "eagle_gather_update.graph": 14,
+        "testLoop.graph": 4,
+        "cont_img_mvp.graph": 45,
+        "test_grpby_gather.graph": 21,
+        "chiles_simple.graph": 22,
+        "Plasma_test.graph": 6,
+    }
+
+    def test_lg_init(self):
+        for lg_name, num_keys in self.lg_names.items():
+            fp = path_utils.get_lg_fpath("logical_graphs", lg_name)
+            lg = LG(fp, ssid=TEST_SSID)
+            self.assertEqual(num_keys,
+                             len(lg._done_dict.keys()),
+                             f"Incorrect number of elements when constructing LG "
+                             f"object using: {lg_name}")
+
+
+def _calc_num_drops(drop_values):
     """
-    Verify the correct data is stored in the LGNode class when a new LGNode is created
-    during the LG constructor.
+    Get the number of drops created during the lgn_to_pgn method.
 
-    This test is here to demonstrate what information is queryable at initial load time,
-    which will facilitate backwards-compatibility testing if changes are made to the
-    internal structure of the class.
+    The drops are stored in dictionaries of the original LG node, so we
+    iterate through them and get the number of Physical Graph Nodes for every one of
+    the Logical Graph nodes.
     """
-    graph_name = "eagle_gather_update.graph"
-    graph_information = {"num_keys": 14}
-    lg = LG(path_utils.get_lg_fpath('logical_graphs', graph_name), ssid="TEST")
-    # Isolate (first) scatter node
-    scatterKey = 0
-    for key, lgn in lg._done_dict.items():
-        if lgn.is_scatter:
-            scatterKey = lgn.id
-            break
-    # Confirm releationship with children
-
-    # TODO Get information on what the structure of the stored data looks like here
-    # This includes the state of the LG class as a whole (e.g. what is in done_dict,
-    # drop_dict, lg_links etc.)
-
-    # TODO We should provide some tests that test things like querying a given LGNode's
-    # children, or confirm the DOP of a given node. That way when we introduce new
-    # representations, we have something to compare to.
+    return sum([len(drop_list) for drop_list in drop_values])
 
 
 class TestLGNToPGN(unittest.TestCase):
@@ -75,31 +74,37 @@ class TestLGNToPGN(unittest.TestCase):
     Intended as a regression testing to ensure backwards compatiblity with existing
     PGT behaviour.
 
-    TODO Get information on what the structure of the stored data looks like here
-    This includes the state of the LG class as a whole (e.g. what is in done_dict,
-    drop_dict, lg_links etc.)
-    """
+    # Note that currently, LGN to PGN creates all the physical graph nodes,
+    but doesn't get rid of construct nodes (these are removed by the unroll_to_tpl)
 
-    def test_simple_lgn_to_pgn(self):
-        """
-        Base-case Hello World application
-        """
+    """
 
     def test_loop_graph(self):
         """
         More complex looping graph
         """
         graph_name = "testLoop.graph"
-        graph_information = {"num_keys": 14}
+        graph_information = {"num_keys": 11}
         lg = LG(path_utils.get_lg_fpath('logical_graphs', graph_name), ssid="TEST")
-        # TODO make this pick specifically the Loop node
         for lgn in lg._start_list:
             lg.lgn_to_pgn(lgn)
 
+        self.assertEqual(graph_information["num_keys"],
+                         _calc_num_drops(lg._drop_dict.values()))
+
     def test_scatter_graph_graph(self):
         """
-        Add scatter and gather constructs
+        Test scatter gather constructs
         """
+
+        graph_name = "eagle_gather_update.graph"
+        graph_information = {"num_keys": 31}
+        lg = LG(path_utils.get_lg_fpath('logical_graphs', graph_name), ssid="TEST")
+        for lgn in lg._start_list:
+            lg.lgn_to_pgn(lgn)
+
+        self.assertEqual(graph_information["num_keys"],
+                         _calc_num_drops(lg._drop_dict.values()))
 
     def test_non_recursive(self):
         """
@@ -110,64 +115,43 @@ class TestLGNToPGN(unittest.TestCase):
                           ssid="TEST")
         for lgn in lg_recursive._start_list:
             lg_recursive.lgn_to_pgn(lgn)
-        lg_non_recursive = LG(path_utils.get_lg_fpath("logical_graphs", graph_name), ssid="TEST")
+        lg_non_recursive = LG(path_utils.get_lg_fpath("logical_graphs", graph_name),
+                              ssid="TEST")
         for lgn in lg_non_recursive._start_list:
             lg_non_recursive.lgn_to_pgn(lgn, recursive=False)
-        for key in lg_recursive._drop_dict:
-            self.assertEqual(lg_recursive._drop_dict[key],
-                             lg_non_recursive._drop_dict[key])
+
+        expected_test_loop_drops = 11
+        self.assertEqual(
+            expected_test_loop_drops,
+            _calc_num_drops(lg_recursive._drop_dict.values())
+        )
+        self.assertEqual(
+            expected_test_loop_drops,
+            _calc_num_drops(lg_non_recursive._drop_dict.values())
+        )
 
         graph_name = "eagle_gather_update.graph"
-
-        lg_recursive = LG(path_utils.get_lg_fpath("logical_graphs", graph_name), ssid="TEST")
+        lg_recursive = LG(path_utils.get_lg_fpath("logical_graphs", graph_name),
+                          ssid="TEST")
         for lgn in lg_recursive._start_list:
             lg_recursive.lgn_to_pgn(lgn)
-        lg_non_recursive = LG(path_utils.get_lg_fpath("logical_graphs", graph_name), ssid="TEST")
+        lg_non_recursive = LG(path_utils.get_lg_fpath("logical_graphs", graph_name),
+                              ssid="TEST")
         for lgn in lg_non_recursive._start_list:
             lg_non_recursive.lgn_to_pgn(lgn, recursive=False)
 
-        with open('lg_recursive.json', 'w') as fp:
-            json.dump(lg_recursive._drop_dict, fp, indent=2)
-        with open('lg_non_recursive.json', 'w') as fp:
-            json.dump(lg_non_recursive._drop_dict, fp, indent=2)
+        self.assertEqual(len(lg_recursive._drop_dict), len(lg_non_recursive._drop_dict))
 
-        for key in lg_recursive._drop_dict:
-            self.assertEqual(lg_recursive._drop_dict[key],
-                             lg_non_recursive._drop_dict[key])
+        expected_test_gather_drops = 31
 
-
-class TestLGInit(unittest.TestCase):
-    """
-    Verify the correct data is stored in the LG class when initially loading a logical
-    graph from file.
-
-    This test is here to demonstrate what information is queryable at initial load time,
-    which will facilitate backwards-compatibility testing if changes are made to the
-    internal structure of the class.
-
-    Tests in this class will include verifying:
-    - What construct something is
-    - What children of that construct are
-    - The relationships between LGNodes (i.e. links/edges)
-
-    """
-
-    def test_initial_relationships(self):
-        """
-        Confirm that self._lgn_list and self._lg_links are correct
-        """
-
-    lg_names = {
-        "HelloWorld_simple.graph": {"num_keys": 2},
-        "eagle_gather_empty_update.graph": {"num_keys": 11},
-        "eagle_gather_simple_update.graph": {"num_keys": 18},
-        "eagle_gather_update.graph": {"num_keys": 14},
-        "testLoop.graph": {"num_keys": 4},
-        "cont_img_mvp.graph": {"num_keys": 45},
-        "test_grpby_gather.graph": {"num_keys": 21},
-        "chiles_simple.graph": {"num_keys": 22},
-        "Plasma_test.graph": {"num_keys": 6},
-    }
+        self.assertEqual(
+            expected_test_gather_drops,
+            _calc_num_drops(lg_recursive._drop_dict.values())
+        )
+        self.assertEqual(
+            expected_test_gather_drops,
+            _calc_num_drops(lg_non_recursive._drop_dict.values())
+        )
 
 
 class TestLGNodeLoading(unittest.TestCase):
@@ -193,7 +177,25 @@ class TestLGUnroll(unittest.TestCase):
     generated using the code they are testing. If the LG class and it's methods change
     in the future, test data may need to be re-generated (provided test failures are
     caused by known-breaking changes, as opposed to legitimate bugs!).
+
+    We no longer compare directly the output, as this causes errors with UIDs/OID
+    conflicts. What we care about in this scenario is that twe have the correct nu
+
     """
+
+    lg_names = {
+        "HelloWorld_simple.graph": {"nodes": 2, "edges": 1},
+        "eagle_gather_empty_update.graph": {"nodes": 22, "edges": 24},
+        "eagle_gather_simple_update.graph": {"nodes": 42, "edges": 55},
+        "eagle_gather_update.graph": {"nodes": 29, "edges": 30},
+        "testLoop.graph": {"nodes": 11, "edges": 10},
+        "cont_img_mvp.graph": {"nodes": 144, "edges": 188},
+        "test_grpby_gather.graph": {"nodes": 15, "edges": 14},
+        "chiles_simple.graph": {"nodes": 22, "edges": 21},
+        "Plasma_test.graph": {"nodes": 6, "edges": 5},
+        "SharedMemoryTest_update.graph": {"nodes": 8, "edges": 7},
+        # "simpleMKN_update.graph", # Currently broken
+    }
 
     def test_lg_unroll(self):
         """
@@ -201,34 +203,38 @@ class TestLGUnroll(unittest.TestCase):
 
         lg_names = { "logical_graph_file.graph": num_keys_in_drop_list, ...}
         """
-
-        lg_names = {
-            "HelloWorld_simple.graph": 2,
-            "eagle_gather_empty_update.graph": 11,
-            "eagle_gather_simple_update.graph": 18,
-            "eagle_gather_update.graph": 14,
-            "testLoop.graph": 4,
-            "cont_img_mvp.graph": 45,
-            "test_grpby_gather.graph": 21,
-            "chiles_simple.graph": 22,
-            "Plasma_test.graph": 6,
-        }
-
-        for lg_name, num_keys in lg_names.items():
+        # TODO These are number of logical graph nodes! Make this exclusive to LG init
+        for lg_name, num_keys in self.lg_names.items():
             fp = path_utils.get_lg_fpath("logical_graphs", lg_name)
             lg = LG(fp, ssid=TEST_SSID)
-            self.assertEqual(num_keys,
-                             len(lg._done_dict.keys()),
-                             f"Incorrect number of elements when constructing LG "
-                             f"object using: {lg_name}")
 
             drop_list = lg.unroll_to_tpl()
             with open(path_utils.get_lg_fpath('pickle', lg_name), 'rb') as pk_fp:
                 test_unroll = pickle.load(pk_fp)
 
-            self.assertListEqual(test_unroll,
-                                 drop_list,
-                                 f"unroll_to_tpl failed for: {lg_name}")
+            # It is worth mentioning that we do not get an accurate number of links
+            # from the LG, as it is not tracked after the initial graph_loading.
+            self.assertEqual(len(test_unroll), len(drop_list))
+            self.assertEqual(num_keys['nodes'], len(drop_list))
+
+            # Confirm number of output/consumers and inputs/producers are the same
+            for i, drop in enumerate(drop_list):
+                if 'outputs' in drop:
+                    expected = test_unroll[i]['outputs']
+                    actual = drop['outputs']
+                    self.assertEqual(len(expected), len(actual))
+                if 'inputs' in drop:
+                    expected = test_unroll[i]['inputs']
+                    actual = drop['inputs']
+                    self.assertEqual(len(expected), len(actual))
+                if 'producers' in drop:
+                    expected = test_unroll[i]['producers']
+                    actual = drop['producers']
+                    self.assertEqual(len(expected), len(actual))
+                if 'consumers' in drop:
+                    expected = test_unroll[i]['consumers']
+                    actual = drop['consumers']
+                    self.assertEqual(len(expected), len(actual))
 
     def test_lg_unroll_sharedmemory(self):
         """
